requirements.txt
numpy==2.4.0
pandas==2.3.3
yfinance==1.0
langchain==1.2.0
langgraph==1.0.5

requests==2.32.5
httpx==0.28.1
cffi==2.0.0
python-dotenv==1.0.1